// src/app/types.ts
export type Role = "admin" | "volunteer";
export type Screen =
  | "splash"
  | "login"
  | "register"
  | "ngo_register"
  | "admin_home"
  | "admin_notifications"
  | "admin_events"
  | "admin_volunteers"
  | "admin_donations"
  | "admin_analytics"
  | "volunteer_home"
  | "ngo_details"
  | "ngo_profile"
  | "participation_history";